feat: support pluggable runtime config
* export runtimeConfig.runtime.ts to manually set the client to be compatible with specific runtime
* get rid of rollup, instead using browser property to swap runtime config
* add endpoint to the serializer utilities and insert it when building a request
