+++ clients/node/client-rds-data-node/.npmignore
@@ -1,6 +1,4 @@
 /coverage/
 /docs/
-*.ts
-!*.d.ts
 tsconfig.test.json
 *.tsbuildinfo
