--- clients/node/client-rds-data-node/index.browser.ts
@@ -1,4 +1,0 @@
-
-export * from "./RDSDataClient";
-export * from "./RDSDataConfiguration";
-export * from "./commands/ExecuteStatementCommand";
