+++ clients/node/client-rds-data-node/package.json
@@ -14,17 +14,18 @@
     "test": "exit 0",
     "smoke-test": "npm run pretest && node ./test/smoke/index.spec.js",
     "build:es": "tsc -p tsconfig.es.json",
-    "build:browser": "rollup -c",
-    "build": "yarn pretest && yarn build:es && yarn build:browser"
+    "build": "yarn pretest && yarn build:es"
   },
-  "main": "./index.js",
+  "main": "./dist/cjs/index.js",
   "types": "./types/index.d.ts",
   "author": {
     "name": "AWS SDK for JavaScript Team",
     "url": "https://aws.amazon.com/javascript/"
   },
   "module": "./dist/es/index.js",
-  "browser": "./dist/browser/rds-data-browser.js",
+  "browser": {
+    "./runtimeConfig": "./runtimeConfig.browser"
+  },
   "sideEffects": false,
   "license": "Apache-2.0",
   "dependencies": {
@@ -47,6 +48,7 @@
     "@aws-sdk/signature-v4": "^0.1.0-preview.7",
     "@aws-sdk/signing-middleware": "^0.1.0-preview.7",
     "@aws-sdk/stream-collector-node": "^0.1.0-preview.6",
+    "@aws-sdk/stream-collector-browser": "^0.1.0-preview.5",
     "@aws-sdk/types": "^0.1.0-preview.5",
     "@aws-sdk/url-parser-browser": "^0.1.0-preview.5",
     "@aws-sdk/url-parser-node": "^0.1.0-preview.5",
@@ -64,10 +66,6 @@
     "@aws-sdk/client-documentation-generator": "^0.1.0-preview.3",
     "@types/node": "^10.0.0",
     "rimraf": "^2.6.2",
-    "rollup-plugin-json": "^4.0.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-sourcemaps": "^0.4.2",
     "typedoc": "^0.14.2",
     "typescript": "^3.7.0-dev.20190926"
   }
