+++ clients/node/client-rds-data-node/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "alwaysStrict": true,
     "target": "es2017",
     "module": "commonjs",
     "declaration": true,
@@ -10,7 +11,9 @@
     "noEmitHelpers": true,
     "incremental": true,
     "resolveJsonModule": true,
-    "declarationDir": "./types"
+    "noUnusedLocals": true,
+    "declarationDir": "./types",
+    "outDir": "dist/cjs"
   },
   "typedocOptions": {
     "exclude": "**/node_modules/**",
