+++ packages/config-resolver/src/index.ts
@@ -1,1 +1,3 @@
-export * from "./components";
+export * from "./EndpointsConfig";
+export * from "./RegionConfig";
+export * from "./ProtocolConfig";
