+++ packages/url-parser-browser/src/index.ts
@@ -1,7 +1,7 @@
 import { parseQueryString } from "@aws-sdk/querystring-parser";
-import { HttpEndpoint, QueryParameterBag, UrlParser } from "@aws-sdk/types";
+import { Endpoint, QueryParameterBag, UrlParser } from "@aws-sdk/types";
 
-export const parseUrl: UrlParser = (url: string): HttpEndpoint => {
+export const parseUrl: UrlParser = (url: string): Endpoint => {
   const { hostname, pathname, port, protocol, search } = new URL(url);
 
   let query: QueryParameterBag | undefined;
