+++ packages/url-parser-universal/src/index.ts
@@ -1,4 +1,4 @@
-import { HttpEndpoint, QueryParameterBag, UrlParser } from "@aws-sdk/types";
+import { UrlParser } from "@aws-sdk/types";
 import { parseUrl as browserUrlParser } from "@aws-sdk/url-parser-browser";
 import { parseUrl as nodeUrlParser } from "@aws-sdk/url-parser-node";
